requirements.txt
requests
tabulate
dnspython
python-whois
